add tests - IV
